Owner on Pokemon

+++ comments/views.py
@@ -1,14 +1,18 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status
-from rest_framework.exceptions import NotFound
+from rest_framework.exceptions import NotFound, PermissionDenied
+from rest_framework.permissions import IsAuthenticated
 
 from .serializers.common import CommentSerializer
 from .models import Comment
 
 class CommentListView(APIView):
 
+    permission_classes = (IsAuthenticated, )
+
     def post(self, request):
+        request.data['owner'] = request.user.id
         comment_to_create = CommentSerializer(data=request.data)
         if comment_to_create.is_valid():
             comment_to_create.save()
@@ -18,9 +22,13 @@ class CommentListView(APIView):
 
 class CommentDetailView(APIView):
 
-    def delete(self, _request, pk):
+    permission_classes = (IsAuthenticated, )
+
+    def delete(self, request, pk):
         try:
             comment_to_delete = Comment.objects.get(pk=pk)
+            if comment_to_delete.owner.id != request.user.id:
+                raise PermissionDenied()
             comment_to_delete.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
         except Comment.DoesNotExist:
